requirements.txt
requests~=2.27.1
click~=8.0.3
beautifulsoup4~=4.10.0
python-telegram-bot==13.10
pandas
pymongo
matplotlib